requirements.txt
fastapi 
uvicorn
requests
authlib
azure-cosmos
python-dotenv